requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
packaging==24.2
postgres==4.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
sqlparse==0.5.3
tzdata==2024.2
whitenoise==6.8.2